Rename namespace helpers file, break up vzconfig utilities into different files
