+++ platform-operator/internal/k8s/namespace/namespace_test.go
